Switch to latest uniffi
